fix(mod): derive the tick slew coefficient in double
The interval-relative law made tau proportional to the lane cycle, so at
the slow end of the panel (RATE 0, PACE 0, TIDE 0) k = 1/(tau*sr) reaches
~1e-8. The tick twin's coefficient is 1 - (1-k)^96, and in float `1.f - k`
rounds to exactly 1.0f below half an ulp: the coefficient quantises, and
at the extreme it is exactly zero -- a one-pole that never moves again.
The old absolute law capped tau at 0.5 s and never went near this, so the
rework introduced it.

Driving ModLane directly at 0.0003125 Hz (what RATE 0 + PACE 0 hand
LANE_SIZE) the lane freezes outright: p2p 0.000000000 over 60 s at SMOOTH
0.50, 0.70 and 1.00, and one single coefficient shared by every knob
position from 0.15 to 0.40. Through the whole instrument the outright
freeze is NOT reachable -- other per-tick motion keeps p2p off zero -- so
G6 gates what is reachable there: that SMOOTH still resolves at all.

G6's RED is the quantisation itself. Against the float coefficient,
SOURCE's p2p at SMOOTH 0.60 and 1.00 are the same number to the last
digit (0.00240338 vs 0.00240338) and the CHECK fires; in double the
three setpoints are strictly decreasing (0.00398457 / 0.00293958 /
0.00272623).

The per-sample slew is deliberately left in float: k itself is perfectly
representable, and it was measured tracking the analytic settling curve
at these same tau values. Only the tick half needed the precision.

wave_formant_sweep is re-cut because the coefficient reaches every render.
This one is rounding, not sound: RMS identical to 0.01 dB against the
render the owner accepted, difference signal 59.2 dB below it, worst
single sample 17 LSB of 32768. The accepted ALT->NEU change sat 4.1 dB
below the render, for scale.

+++ engine/mod/lane.cpp
@@ -413,9 +413,27 @@ void ModLane::_update_slew() {
     // exactly. If that formula changes, this must change with it -- these
     // two are a matched pair, not independent code, and this tick twin would
     // otherwise silently diverge from process()'s slew.
-    float k = 1.f / (t * _sr);
-    if (k > 1.f) k = 1.f;
-    _slew_tick.set_coef(1.f - std::pow(1.f - k, static_cast<float>(kTickInterval)));
+    // In DOUBLE, and not as a style preference. Under the old absolute law tau
+    // was capped at 0.5 s, so k never fell below 4e-5. It is now proportional
+    // to the cycle, and at the slow end of the panel k reaches ~1e-8. In float
+    // `1.f - k` rounds to exactly 1.0f below half an ulp (k < 2.98e-8), so the
+    // power is 1 and the coefficient QUANTISES -- and at the extreme reaches
+    // exactly zero, which is a one-pole that never moves again.
+    //
+    // Measured on ModLane driven directly at 0.0003125 Hz (what RATE 0 + PACE 0
+    // hand LANE_SIZE), 60 s: p2p 0.000000000 at SMOOTH 0.50, 0.70 and 1.00, and
+    // one single coefficient shared by every knob position from 0.15 to 0.40.
+    // Through the whole instrument the outright freeze is NOT reachable -- other
+    // per-tick motion keeps p2p off zero -- but the quantisation is, and
+    // tests/test_smooth_law.cpp's G6 gates it there.
+    //
+    // The per-sample _slew above is NOT affected: k itself is perfectly
+    // representable, and it was measured tracking the analytic settling curve
+    // at these same tau values. Only this half of the pair needs the precision.
+    double k = 1.0 / (double(t) * double(_sr));
+    if (k > 1.0) k = 1.0;
+    _slew_tick.set_coef(static_cast<float>(
+        1.0 - std::pow(1.0 - k, static_cast<double>(kTickInterval))));
 }
 
 void ModLane::kick(float dphase, float dshape) {
